+++ src/group/stabchain/builder/random/random_ift.rs
@@ -458,7 +458,7 @@ where
 
 // Functions used for compatability reasons
 /// Generate a word representation of a random subproduct of the given generators.
-pub fn random_subproduct_word_subset<R, P>(rng: &mut R, gens: &[P], k: usize) -> Vec<P>
+fn random_subproduct_word_subset<R, P>(rng: &mut R, gens: &[P], k: usize) -> Vec<P>
 where
     P: Permutation,
     R: Rng,
@@ -471,7 +471,7 @@ where
 }
 
 /// Generate random subproduct of the given generators.
-pub fn random_subproduct_word_full<T, P>(rng: &mut T, gens: &[P]) -> Vec<P>
+fn random_subproduct_word_full<T, P>(rng: &mut T, gens: &[P]) -> Vec<P>
 where
     P: Permutation,
     T: Rng,
@@ -480,7 +480,7 @@ where
 }
 
 /// Apply a point to permutations stored as a word.
-pub fn apply_permutation_word<'a, P, A>(
+fn apply_permutation_word<'a, P, A>(
     perm_word: impl IntoIterator<Item = &'a P>,
     x: A::OrbitT,
     strat: &A,
@@ -495,10 +495,41 @@ where
 }
 
 /// Convert from a permutation stored as a word, into a single permutation.
-pub fn collapse_perm_word<'a, P>(p: impl IntoIterator<Item = &'a P>) -> P
+fn collapse_perm_word<'a, P>(p: impl IntoIterator<Item = &'a P>) -> P
 where
     P: 'a + Permutation,
 {
     p.into_iter()
         .fold(Permutation::id(), |accum, perm| accum.multiply(perm))
+}
+
+mod tests {
+
+    ///Test that applying a permutation as a word gives the same image as collapsing that permutation.
+    #[test]
+    fn test_apply_permutation_word() {
+        use super::apply_permutation_word;
+        use super::collapse_perm_word;
+        use crate::perm::actions::SimpleApplication;
+        use crate::perm::export::CyclePermutation;
+        use crate::perm::impls::standard::StandardPermutation;
+        use crate::perm::Permutation;
+        //Test an empty word.
+        let empty_word = vec![];
+        let strat = SimpleApplication::default();
+        assert_eq!(3, apply_permutation_word(&empty_word, 3, &strat));
+        let perm_word: Vec<StandardPermutation> = vec![
+            CyclePermutation::single_cycle(&[1, 2, 4]).into(),
+            CyclePermutation::single_cycle(&[3, 5, 8]).into(),
+            CyclePermutation::single_cycle(&[7, 9]).into(),
+            CyclePermutation::single_cycle(&[1, 5, 6, 9]).into(),
+        ];
+        let collapsed_word = collapse_perm_word(&perm_word);
+        for i in 0..9 {
+            assert_eq!(
+                collapsed_word.apply(i),
+                apply_permutation_word(&perm_word, i, &strat)
+            );
+        }
+    }
 }
